+++ das_backend/src/main/java/ch/sbb/das/backend/preload/infrastructure/xml/SferaMessagingConfig.java
@@ -1,6 +1,6 @@
 package ch.sbb.das.backend.preload.infrastructure.xml;
 
-import ch.sbb.das.backend.preload.sfera.model.v0300.ObjectFactory;
+import ch.sbb.das.backend.preload.sfera.model.v0400.ObjectFactory;
 import jakarta.xml.bind.Marshaller;
 import java.util.Map;
 import org.springframework.context.annotation.Bean;
@@ -11,13 +11,13 @@ import org.springframework.oxm.jaxb.Jaxb2Marshaller;
 @Configuration
 public class SferaMessagingConfig {
 
-    public static final String SFERA_V3_XSD = "SFERA_v3.00.xsd";
+    public static final String SFERA_V4_XSD = "SFERA_4.00.xsd";
 
     @Bean
     public Jaxb2Marshaller jaxb2Marshaller() {
         var marshaller = new Jaxb2Marshaller();
         marshaller.setContextPath(ObjectFactory.class.getPackageName());
-        marshaller.setSchemas(new ClassPathResource(SFERA_V3_XSD));
+        marshaller.setSchemas(new ClassPathResource(SFERA_V4_XSD));
         marshaller.setMarshallerProperties(Map.of(Marshaller.JAXB_FRAGMENT, true)); // suppress xml prolog
         return marshaller;
     }