+++ das_client/app/lib/pages/journey/journey_screen/notification/notification_space.dart
@@ -7,6 +7,7 @@ import 'package:app/pages/journey/journey_screen/notification/widgets/departure_
 import 'package:app/pages/journey/journey_screen/notification/widgets/disturbance_notification.dart';
 import 'package:app/pages/journey/journey_screen/notification/widgets/koa_notification.dart';
 import 'package:app/pages/journey/journey_screen/notification/widgets/maneuver_notification.dart';
+import 'package:app/pages/journey/journey_screen/notification/widgets/reauthentication_required_notification.dart';
 import 'package:app/pages/journey/journey_screen/notification/widgets/replacement_series_notification.dart';
 import 'package:app/pages/journey/journey_screen/notification/widgets/suspicious_segment_notification.dart';
 import 'package:app/pages/journey/journey_screen/view_model/notification_priority_view_model.dart';
@@ -74,8 +75,7 @@ extension _WidgetNotificationTypeX on NotificationType {
   Widget toWidget() {
     return switch (this) {
       .illegalSegmentNoReplacement => ReplacementSeriesNotification(),
-      .koaWait => KoaNotification(),
-      .koaWaitCancelled => KoaNotification(),
+      .koa => KoaNotification(),
       .newBrakeLoadSlip => BrakeLoadSlipNotification(),
       .maneuverMode => ManeuverNotification(),
       .disturbance => DisturbanceNotification(),
@@ -83,6 +83,7 @@ extension _WidgetNotificationTypeX on NotificationType {
       .departureDispatch => DepartureDispatchNotification(),
       .illegalSegmentWithReplacement => ReplacementSeriesNotification(),
       .suspiciousSegment => SuspiciousSegmentNotification(),
+      .reauthenticationRequired => ReauthenticationRequiredNotification(),
     };
   }
 }