+++ das_client/app/lib/pages/journey/journey_screen/notification/notification_type.dart
@@ -1,8 +1,7 @@
 // MUST be ordered according to priority - see https://github.com/SchweizerischeBundesbahnen/DAS/issues/1402
 enum NotificationType {
   illegalSegmentNoReplacement,
-  koaWait,
-  koaWaitCancelled,
+  koa,
   newBrakeLoadSlip,
   maneuverMode,
   disturbance,
@@ -10,4 +9,5 @@ enum NotificationType {
   departureDispatch,
   illegalSegmentWithReplacement,
   suspiciousSegment,
+  reauthenticationRequired,
 }