+++ das_client/app/lib/pages/journey/journey_screen/view_model/notification_priority_view_model.dart
@@ -55,6 +55,9 @@ class NotificationPriorityQueueViewModel extends JourneyAwareViewModel {
   }
 
   void insert({required NotificationType type, VoidCallback? callback}) {
+    if (!_activeNotifications.contains(type)) {
+      _logger.fine('Adding notification of type $type');
+    }
     _activeNotifications.add(type);
     _notificationTypeToCallback[type] = callback;
     _emitMaximumTwoPrioritizedNotifications();
@@ -62,6 +65,9 @@ class NotificationPriorityQueueViewModel extends JourneyAwareViewModel {
   }
 
   void remove({required NotificationType type}) {
+    if (_activeNotifications.contains(type)) {
+      _logger.fine('Removing notification of type $type');
+    }
     _activeNotifications.remove(type);
     _emitMaximumTwoPrioritizedNotifications();
     _callCallbacks();