+++ das_client/auth/lib/src/azure_authenticator.dart
@@ -2,33 +2,45 @@ import 'dart:async';
 
 import 'package:auth/component.dart';
 import 'package:auth/src/oidc_client_provider.dart';
+import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:logging/logging.dart';
+import 'package:rxdart/rxdart.dart';
 import 'package:sbb_oidc/sbb_oidc.dart';
 
 final _log = Logger('AzureAuthenticator');
 
 class AzureAuthenticator implements Authenticator {
-  AzureAuthenticator({
-    required this.config,
-    this.oidcClientFactory = const SBBOidcClientFactory(),
-  });
+  static const _offlineTokenValidityDuration = Duration(days: 1);
 
-  final AuthenticatorConfig config;
-  final OidcClientFactory oidcClientFactory;
+  AzureAuthenticator({
+    required AuthenticatorConfig config,
+    OidcClientFactory oidcClientFactory = const SBBOidcClientFactory(),
+    FlutterSecureStorage storage = const FlutterSecureStorage(
+      iOptions: IOSOptions(accountName: 'auth'),
+      aOptions: AndroidOptions(sharedPreferencesName: 'auth'),
+    ),
+  }) : _config = config,
+       _oidcClientFactory = oidcClientFactory,
+       _storage = storage;
+
+  final AuthenticatorConfig _config;
+  final OidcClientFactory _oidcClientFactory;
   late final OidcClient _oidcClient;
-  bool isInitialized = false;
+  final FlutterSecureStorage _storage;
+  bool _isInitialized = false;
+  final _reauthenticationRequiredSubject = BehaviorSubject.seeded(false);
 
   Future<void> _init() async {
-    if (isInitialized) return;
+    if (_isInitialized) return;
     _log.fine('Initialize AzureAuthenticator');
     try {
-      _oidcClient = await oidcClientFactory.createClient(
-        discoveryUrl: config.discoveryUrl,
-        clientId: config.clientId,
-        redirectUrl: config.redirectUrl,
-        postLogoutRedirectUrl: config.postLogoutRedirectUrl,
+      _oidcClient = await _oidcClientFactory.createClient(
+        discoveryUrl: _config.discoveryUrl,
+        clientId: _config.clientId,
+        redirectUrl: _config.redirectUrl,
+        postLogoutRedirectUrl: _config.postLogoutRedirectUrl,
       );
-      isInitialized = true;
+      _isInitialized = true;
     } catch (e, s) {
       _log.severe('AzureAuthenticator Initialization failed', e, s);
       rethrow;
@@ -37,50 +49,101 @@ class AzureAuthenticator implements Authenticator {
 
   @override
   Future<bool> get isAuthenticated async {
-    await _init();
     try {
       await token();
       return true;
     } catch (e) {
-      // Delete all existing tokens. They are invalid and might cause issues.
-      await logout();
+      if (_isInitialized) {
+        // Delete all existing tokens. They are invalid and might cause issues.
+        await logout();
+      }
       return false;
     }
   }
 
   @override
   Future<OidcToken> login({String? tokenId}) async {
-    await _init();
-    TokenSpec? tokenSpec = config.tokenSpecs.getById(tokenId);
-    tokenSpec ??= config.tokenSpecs.all.first;
-    final token = await _oidcClient.login(
-      scopes: tokenSpec.scopes,
-      prompt: LoginPrompt.selectAccount,
-    );
+    try {
+      await _init();
+      TokenSpec? tokenSpec = _config.tokenSpecs.getById(tokenId);
+      tokenSpec ??= _config.tokenSpecs.all.first;
+      final token = await _oidcClient.login(
+        scopes: tokenSpec.scopes,
+        prompt: LoginPrompt.selectAccount,
+      );
+      _validateToken(token);
+      _updateReauthenticationRequiredState(false);
+      _writeOfflineToken(token: token, tokenId: tokenId);
 
-    _validateToken(token);
+      return token;
+    } catch (e) {
+      if (!_isNetworkError(e)) {
+        _log.info('Failed to retrieve new token on login, reauthentication required', e);
+        _updateReauthenticationRequiredState(true);
+      }
 
-    return token;
+      final token = await _loadOfflineToken(tokenId: tokenId);
+      if (token != null) {
+        _validateToken(token);
+        _log.info('Login failed, but found valid offline token. Using it as fallback.');
+        return token;
+      }
+      rethrow;
+    }
   }
 
   @override
   Future<OidcToken> token({String? tokenId, bool? forceRefresh}) async {
-    await _init();
-    final tokenSpec = config.tokenSpecs.getById(tokenId);
-    if (tokenSpec == null) {
-      throw ArgumentError.value(tokenId, 'tokenId', 'Unknown token id.');
+    try {
+      await _init();
+      final tokenSpec = _config.tokenSpecs.getById(tokenId);
+      if (tokenSpec == null) {
+        throw ArgumentError.value(tokenId, 'tokenId', 'Unknown token id.');
+      }
+
+      final token = await _oidcClient.getToken(scopes: tokenSpec.scopes, forceRefresh: forceRefresh ?? false);
+      _validateToken(token);
+      _updateReauthenticationRequiredState(false);
+      _writeOfflineToken(token: token, tokenId: tokenId);
+
+      return token;
+    } catch (e) {
+      if (!_isNetworkError(e)) {
+        _log.info('Failed to retrieve new token, reauthentication required', e);
+        _updateReauthenticationRequiredState(true);
+      }
+
+      final token = await _loadOfflineToken(tokenId: tokenId);
+      if (token != null) {
+        _validateToken(token);
+        _log.info('Found valid offline token. Using it as fallback.');
+        return token;
+      }
+      rethrow;
     }
+  }
 
-    final token = await _oidcClient.getToken(scopes: tokenSpec.scopes, forceRefresh: forceRefresh ?? false);
-    _validateToken(token);
+  bool _isNetworkError(dynamic e) {
+    if (e is NetworkException) return true;
 
-    return token;
+    final message = e.toString();
+    if (message.contains('Connection error')) return true;
+
+    return false;
   }
 
   @override
   Future<User> user({String? tokenId}) async {
-    await _init();
-    final oidcToken = await token(tokenId: tokenId);
+    OidcToken? oidcToken;
+    try {
+      await _init();
+      oidcToken = await token(tokenId: tokenId);
+    } catch (e) {
+      oidcToken = await _loadOfflineToken(tokenId: tokenId);
+      if (oidcToken == null) {
+        rethrow;
+      }
+    }
     final idToken = JsonWebToken.decode(oidcToken.idToken);
     final userId = idToken.payload['preferred_username'] as String;
     final roles = idToken.payload['roles'] as List<dynamic>? ?? [];
@@ -95,23 +158,59 @@ class AzureAuthenticator implements Authenticator {
     );
   }
 
+  void _updateReauthenticationRequiredState(bool state) {
+    if (_reauthenticationRequiredSubject.value != state) {
+      _log.info('Updating reauthentication required state to $state');
+      _reauthenticationRequiredSubject.add(state);
+    }
+  }
+
   @override
   Future<void> logout() async {
     await _init();
+    await _storage.deleteAll();
     return _oidcClient.logout();
   }
 
   @override
   Future<void> endSession() async {
     await _init();
+    await _storage.deleteAll();
     return _oidcClient.endSession();
   }
 
+  Future<OidcToken?> _loadOfflineToken({String? tokenId}) async {
+    final key = tokenId ?? TokenSpec.defaultTokenId;
+
+    final tokenPayload = await _storage.read(key: key);
+    if (tokenPayload == null) return null;
+
+    final oidcToken = OidcToken.fromJsonString(tokenPayload);
+    if (_isExpiredMoreThanOfflineValidityDuration(oidcToken)) {
+      await _storage.delete(key: key);
+      return null;
+    }
+
+    return oidcToken;
+  }
+
+  bool _isExpiredMoreThanOfflineValidityDuration(OidcToken oidcToken) {
+    return oidcToken.accessTokenExpirationDateTime == null ||
+        oidcToken.accessTokenExpirationDateTime!.isBefore(DateTime.now().subtract(_offlineTokenValidityDuration));
+  }
+
+  Future<void> _writeOfflineToken({required OidcToken token, String? tokenId}) {
+    return _storage.write(key: tokenId ?? TokenSpec.defaultTokenId, value: token.toJsonString());
+  }
+
   void _validateToken(OidcToken token) {
-    if (!token.isIssuedByTenant(config.trustedTenantIds)) {
+    if (!token.isIssuedByTenant(_config.trustedTenantIds)) {
       throw Exception('Token issued by untrusted tenant');
     }
   }
+
+  @override
+  Stream<bool> get reauthenticationRequired => _reauthenticationRequiredSubject.distinct();
 }
 
 extension _OidcTokenExtension on OidcToken {