+++ das_client/sfera/lib/component.dart
@@ -1,3 +1,4 @@
+import 'package:auth/component.dart';
 import 'package:connectivity_x/component.dart';
 import 'package:mqtt/component.dart';
 import 'package:sfera/src/data/local/drift_sfera_local_database_service.dart';
@@ -89,6 +90,8 @@ class SferaComponent {
     required String deviceId,
     required SferaLocalRepo localRepo,
     required ConnectivityManager connectivityManager,
+    required Authenticator authenticator,
+    required String sferaVersion,
   }) {
     final localDatabaseService = DriftSferaLocalDatabaseService.instance;
     return SferaRepoImpl(
@@ -98,6 +101,8 @@ class SferaComponent {
       localRepo: localRepo,
       connectivityManager: connectivityManager,
       deviceId: deviceId,
+      authenticator: authenticator,
+      sferaVersion: sferaVersion,
     );
   }