+++ das_client/sfera/lib/src/data/dto/handshake_acknowledgement_dto.dart
@@ -2,7 +2,7 @@ import 'package:sfera/src/data/dto/das_operating_modes_selected_dto.dart';
 import 'package:sfera/src/data/dto/sfera_xml_element_dto.dart';
 
 class HandshakeAcknowledgementDto extends SferaXmlElementDto {
-  static const String elementType = 'HandshakeAcknowledgement';
+  static const String elementType = 'DAS_HandshakeAcknowledgement';
 
   HandshakeAcknowledgementDto({super.type = elementType, super.attributes, super.children, super.value});