+++ das_client/sfera/lib/src/data/dto/taf_tap_route_table_data_nsp_dto.dart
@@ -1,6 +1,7 @@
 import 'package:sfera/component.dart';
 import 'package:sfera/src/data/dto/route_table_data_relevant_wrapper_dto.dart';
 import 'package:sfera/src/data/dto/taf_tap_location_nsp_dto.dart';
+import 'package:sfera/src/data/parser/parse_utils.dart';
 
 class TafTapRouteTableDataNspDto extends TafTapLocationNspDto {
   static const String groupNameValue = 'routeTableData';
@@ -20,4 +21,8 @@ class TafTapRouteTableDataNspDto extends TafTapLocationNspDto {
       children.whereNspWithName('betweenBrackets').map((it) => bool.tryParse(it.nspValue)).firstOrNull ?? false;
 
   String? get trackGroup => children.whereNspWithName('trackGroup').firstOrNull?.nspValue;
+
+  double? get km1 => ParseUtils.tryParseDouble(children.whereNspWithName('km1').map((it) => it.nspValue).firstOrNull);
+
+  double? get km2 => ParseUtils.tryParseDouble(children.whereNspWithName('km2').map((it) => it.nspValue).firstOrNull);
 }